initial work to implement structured data
