+++ src/routes/blog/[slug]/+page.js
@@ -2,8 +2,8 @@ import { parse } from '$lib/blog.js'
 
 /** @type {import('./$types').PageLoad} */
 export async function load({ params }) {
-  const [_, title] = parse(params.slug)
+  const [date, title] = parse(params.slug)
   const description = ""
 
-  return { title, description }
+  return { title, description, date }
 }
